Unify the touch helpers as well
